+++ packages/preact/src/index.tsx
@@ -1,21 +1,30 @@
+import type { JSX } from "preact";
+
 import { cx } from "@classed/utils";
 import type { CXLike } from "@classed/utils";
 
-type Create = <CX extends CXLike>(options: { cx: CX }) => (element: string, ...classes: Parameters<CX>) => { type: string; props: { class: string } };
+import type { SupportedComponentProps, SupportedElementType, Classes, BaseComponent } from "./types";
 
-const create: Create = ({ cx }: { cx: any }) => {
-  return (element, ...classes) => {
-    return {
-      type: element,
-      props: {
-        class: cx(...classes),
-      },
-    };
+const isSignal = <T extends any>(obj: any): obj is JSX.SignalLike<T> => obj !== null && typeof obj === "object" && obj.brand === Symbol.for("preact-signals");
+
+const maybeSignal = <T extends any = any>(obj: any): obj is T => (isSignal<any>(obj) ? obj.value : obj);
+
+type WithCXValueProps<T, P extends {}> = Omit<P, Classes> & Partial<Record<Classes, JSX.Signalish<T>>>;
+
+type ClassedComponent<ET extends SupportedElementType, CXValue> = {
+  (props: WithCXValueProps<CXValue, SupportedComponentProps<ET>>): JSX.Element;
+} & BaseComponent;
+
+const create = <CX extends CXLike>({ cx }: { cx: CX }) => {
+  type CXValue = Parameters<CX>[number];
+  return <ET extends SupportedElementType>(Element: ET, ...classes: CXValue[]) => {
+    return ((props) => {
+      return <Element {...(props as any)} class={cx(...classes, maybeSignal(props.class) ?? maybeSignal(props.className))} />;
+    }) as ClassedComponent<ET, CXValue>;
   };
 };
 
 const classed = create({ cx });
 
-export type { Create };
 export { create };
 export default classed;