+++ packages/qwik/test/expects.ts
@@ -1,0 +1,19 @@
+import { expect } from "vitest";
+
+export const element = <T extends Element>(el: T) => {
+  expect(el).toBeDefined();
+  return {
+    className(className: string) {
+      expect(el.className).toEqual(className);
+      return this;
+    },
+    tagName(tagName: string) {
+      expect(el.tagName).toEqual(tagName);
+      return this;
+    },
+    textContent(textContent: string) {
+      expect(el.textContent).toEqual(textContent);
+      return this;
+    },
+  };
+};