+++ packages/react/test/index.test.tsx
@@ -1,22 +1,34 @@
-import { describe, expect, it } from "vitest";
+import { describe, it } from "vitest";
 
-import clsx from "clsx";
+import { render, fireEvent } from "@testing-library/react";
 
-import classed, { create } from "./../src";
+import * as expects from "./expects";
 
-describe("test", () => {
-  it("works", () => {
-    const element = classed("div", "foo", "bar");
-    expect(element.type).toEqual("div");
-    expect(element.props.class).toEqual("foo bar");
+import { ClassedButton, ClassedButtonReactive } from "./index.test.utils";
+
+describe("ClassedButton", () => {
+  it("default", () => {
+    const element = render(
+      <ClassedButton data-testid="root" className="extra classes">
+        children
+      </ClassedButton>
+    ).getByTestId("root");
+
+    expects.element(element).tagName("BUTTON").className("button extra classes").textContent("children");
   });
 
-  describe("custom", () => {
-    it("works", () => {
-      const classed = create({ cx: clsx });
-      const element = classed("div", "foo", "bar");
-      expect(element.type).toEqual("div");
-      expect(element.props.class).toEqual("foo bar");
-    });
+  it("reactive", () => {
+    const { getByTestId } = render(<ClassedButtonReactive />);
+
+    let element = getByTestId("reactive");
+
+    expects.element(element).tagName("BUTTON").className("button extra classes");
+
+    fireEvent.click(element);
+
+    expects
+      .element((element = getByTestId("reactive")))
+      .tagName("BUTTON")
+      .className("button extra classes reactive");
   });
 });