+++ packages/utils/test/index.test.ts
@@ -2,12 +2,13 @@ import { describe, expect, it } from "vitest";
 
 import { cx } from "./../src";
 
-describe("test", () => {
+describe("cx", () => {
   it("works", () => {
     expect(cx()).toEqual("");
     expect(cx("foo")).toEqual("foo");
     expect(cx("foo bar")).toEqual("foo bar");
     expect(cx("foo", "bar")).toEqual("foo bar");
     expect(cx(null, "foo", false, "bar", undefined)).toEqual("foo bar");
+    expect(cx(0 as any, "foo", {} as any, "bar", [] as any)).toEqual("foo bar");
   });
 });