+++ packages/qwik/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @classed/qwik
 
+## 0.1.0
+
+### Minor Changes
+
+- prototype packages
+
+### Patch Changes
+
+- Updated dependencies
+  - @classed/utils@0.1.0
+
 ## 0.0.0
 
 ### Major Changes